mongo database populated with 10M documents
