+++ database/postgresSchema.sql
@@ -22,6 +22,6 @@ CREATE TABLE topbunk.listings (
 	num_baths INTEGER
 );
 
-\copy topbunk.listings (room_type, username, room_details, city, city_details, listing_details, guest_access, interaction, other, avatar, num_guests, num_bedrooms, num_beds, num_baths) from 'data.csv' with delimiter as E'\t' CSV;
+\copy topbunk.listings (id, room_type, username, room_details, city, city_details, listing_details, guest_access, interaction, other, avatar, num_guests, num_bedrooms, num_beds, num_baths) from 'data.csv' with delimiter as E'\t' CSV;
 
 ALTER TABLE topbunk.listings ADD PRIMARY KEY (id);
