+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "react-dev": "webpack -d --watch",
     "server-dev": "nodemon server/index.js",
-    "build-csv": "node ./database/stream.js",
+    "build-csv": "node ./database/build.js",
     "seed-postgres": "psql topbunk < ./database/postgresSchema.sql",
     "seed-mongo": "./database/mongoInit.sh"
   },
